refactor: remove implicit error handling and add example on how to handle errors
